Bounties are not used anymore, only Submissions, "Scan" is only an "ArtifactInstance" really
