+++ src/polyswarm_api/types/__init__.py
@@ -3,4 +3,3 @@ from . import hash
 from . import base
 from . import result
 from . import schemas
-from . import scan
