+++ src/polyswarm_api/types/artifact.py
@@ -7,9 +7,85 @@ from .hash import all_hashes, Hash, Hashable
 from io import BytesIO
 import os
 from . import schemas
-from .scan import Bounty, Vote, Assertion
 from . import date
 from .. import const
+
+
+class Submission(BasePSJSONType):
+    SCHEMA = schemas.bounty_schema
+
+    def __init__(self, instance, json, polyswarm=None):
+        super(Submission, self).__init__(json, polyswarm)
+        self.instance = instance
+        self.artifact_type = ArtifactType.from_string(json['artifact_type']) if json.get('artifact_type') else \
+            ArtifactType.FILE
+        self.status = json['status']
+        self.uuid = json.get('uuid')
+        self._permalink = json['permalink'] if json.get('permalink') else None
+        self.failed = self.status == 'Bounty Failed'
+
+    def __str__(self):
+        return "Submission-%s" % self.uuid
+
+
+class PolyScore(BasePSJSONType):
+    SCHEMA = schemas.polyscore_schema
+
+    def __init__(self, json, polyswarm=None):
+        super(PolyScore, self).__init__(json, polyswarm)
+
+        self.scores = json['scores']
+
+    def get_score_by_id(self, instance_id):
+        return self.scores.get(str(instance_id), None)
+
+
+class Assertion(BasePSJSONType):
+    SCHEMA = schemas.assertion_schema
+
+    def __init__(self, scanfile, json, polyswarm=None):
+        super(Assertion, self).__init__(json, polyswarm)
+        self.scanfile = scanfile
+        self.author = json['author']
+
+        # TODO this needs to go away, but we are forced to until we do this resolution server-side
+        # this should be removed after next deploy
+        # we also are forced to modify the json in place so that this information is included in JSON formatting
+        # TODO deprecate engine_name
+        if 'author_name' in json:
+            self.engine_name = json['author_name']
+            self.author_name = json['author_name']
+        elif 'engine_name' in json:
+            self.engine_name = json['engine_name']
+            self.author_name = json['engine_name']
+        else:
+            # TODO deprecate
+            self.engine_name = self.polyswarm._resolve_engine_name(self.author) if self.polyswarm else self.author
+            self.json['engine_name'] = self.engine_name
+            self.author_name = self.engine_name
+
+
+        self.bid = int(json['bid'])
+        self.mask = json['mask']
+        # deal with metadata being a string instead of null
+        self.metadata = json['metadata'] if json['metadata'] else {}
+        self.verdict = json['verdict']
+
+    def __str__(self):
+        return "Assertion-%s: %s" % (self.engine_name, self.verdict)
+
+
+class Vote(BasePSJSONType):
+    SCHEMA = schemas.vote_schema
+
+    def __init__(self, scanfile, json, polyswarm=None):
+        super(Vote, self).__init__(json, polyswarm)
+        self.scanfile = scanfile
+        self.arbiter = json['arbiter']
+        self.vote = json['vote']
+
+    def __str__(self):
+        return "Vote-%s: %s" % (self.arbiter, self.vote)
 
 
 def requires_analysis(func):
@@ -50,7 +126,7 @@ class ArtifactInstance(BasePSJSONType):
         self.bounty_id = json['bounty_id']
         # rename this to make it clearer, and avoid the word 'Result' which has a specific meaning in this project
         # using Scan here instead of Bounty
-        self.bounty = Bounty(self, json['bounty_result'], polyswarm=polyswarm) if json['bounty_result'] else None
+        self.bounty = Submission(self, json['bounty_result'], polyswarm=polyswarm) if json['bounty_result'] else None
         self.community = json['community']
         self.submission_uuid = json['submission_uuid']
         self.submission_guid = self.submission_uuid
@@ -71,6 +147,13 @@ class ArtifactInstance(BasePSJSONType):
 
         if self.ready and polyswarm and polyscore:
             self.fetch_polyscore()
+
+    def __str__(self):
+        return "ArtifactInstance-<%s>" % self.hash
+
+    @property
+    def detections(self):
+        return [a for a in self.assertions if a.mask and a.verdict]
 
     @property
     def polyscore(self):
